Move form_fields file
We are using RDU form fields in more than just the CMS, so it doesn't
make sense to have them in that directory.
